Rename ensemble_calibration to emos_calibration and update all references.
