+++ improver/api/__init__.py
@@ -21,7 +21,7 @@ PROCESSING_MODULES = {
     "ApplyBiasCorrection": "improver.calibration.simple_bias_correction",
     "ApplyDecisionTree": "improver.categorical.decision_tree",
     "ApplyDzRescaling": "improver.calibration.dz_rescaling",
-    "ApplyEMOS": "improver.calibration.ensemble_calibration",
+    "ApplyEMOS": "improver.calibration.emos_calibration",
     "ApplyGriddedLapseRate": "improver.temperature.lapse_rate",
     "ApplyNeighbourhoodProcessingWithAMask": "improver.nbhood.use_nbhood",
     "ApplyOrographicEnhancement": "improver.nowcasting.utilities",
@@ -29,7 +29,7 @@ PROCESSING_MODULES = {
     "ApplyReliabilityCalibration": "improver.calibration.reliability_calibration",
     "BaseNeighbourhoodProcessing": "improver.nbhood.nbhood",
     "CalculateForecastBias": "improver.calibration.simple_bias_correction",
-    "CalibratedForecastDistributionParameters": "improver.calibration.ensemble_calibration",
+    "CalibratedForecastDistributionParameters": "improver.calibration.emos_calibration",
     "ChooseDefaultWeightsLinear": "improver.blending.weights",
     "ChooseDefaultWeightsNonLinear": "improver.blending.weights",
     "ChooseDefaultWeightsTriangular": "improver.blending.weights",
@@ -38,7 +38,7 @@ PROCESSING_MODULES = {
     "CloudTopTemperature": "improver.psychrometric_calculations.cloud_top_temperature",
     "Combine": "improver.cube_combiner",
     "ConstructReliabilityCalibrationTables": "improver.calibration.reliability_calibration",
-    "ContinuousRankedProbabilityScoreMinimisers": "improver.calibration.ensemble_calibration",
+    "ContinuousRankedProbabilityScoreMinimisers": "improver.calibration.emos_calibration",
     "ConvectionRatioFromComponents": "improver.precipitation.convection",
     "ConvertProbabilitiesToPercentiles": "improver.ensemble_copula_coupling.ensemble_copula_coupling",
     "CopyMetadata": "improver.utilities.copy_metadata",
@@ -49,7 +49,7 @@ PROCESSING_MODULES = {
     "DifferenceBetweenAdjacentGridSquares": "improver.utilities.spatial",
     "EnforceConsistentForecasts": "improver.utilities.forecast_reference_enforcement",
     "EnsembleReordering": "improver.ensemble_copula_coupling.ensemble_copula_coupling",
-    "EstimateCoefficientsForEnsembleCalibration": "improver.calibration.ensemble_calibration",
+    "EstimateCoefficientsForEnsembleCalibration": "improver.calibration.emos_calibration",
     "EstimateDzRescaling": "improver.calibration.dz_rescaling",
     "ExpectedValue": "improver.expected_value",
     "ExtendRadarMask": "improver.nowcasting.utilities",
